==> notebooks/p1_external_data_download.ipynb <==
# External Data Downloader

This notebook helps to download all needed the external data sets.

import os
import urllib
from urllib.request import urlretrieve
import zipfile
import requests
import tempfile

# Set the working directory
output_dir = f"../data/raw/"

### 1 - School Location

# URL for the school data
URL_school = "http://www.education.vic.gov.au/Documents/about/research/datavic/"

# 2018-2019 school data
URL_school_18_19 = f"{URL_school}dv279-schoollocations2019"

# Generate output location and filename
output = f"{output_dir}school_location_18_19.csv"

# Urlretrieve the file
urlretrieve(f'{URL_school_18_19}.csv', output) 

# 2020 school data
URL_school_20 = f"{URL_school}dv296-schoollocations2020"

# generate output location and filename
output = f"{output_dir}school_location_20.csv"

# Urlretrieve the file
urlretrieve(f'{URL_school_20}.csv', output) 

# 2021 school data
URL_school_21 = f"{URL_school}dv309_schoollocations2021"

# Generate output location and filename
output = f"{output_dir}school_location_21.csv"

# Urlretrieve the file
urlretrieve(f'{URL_school_21}.csv', output) 

# 2022 school data
URL_school_22 = f"{URL_school}dv331_schoollocations2022"

# Generate output location and filename
output = f"{output_dir}school_location_22.csv"

# download
urlretrieve(f'{URL_school_22}.csv', output) 

### 2 - Population data

# URL of data of population
URL_population = "https://www.abs.gov.au/statistics/people/population/regional-population/2021/32180DS0001_2001-21"

# Generate output location and filename
output = f"{output_dir}population.xlsx"

# Urlretrieve the file
urlretrieve(f'{URL_population}.xlsx', output) 

### 3 - Train station

# URL
URL_train = "https://vicroadsopendatastorehouse.vicroads.vic.gov.au/opendata/Public_Transport/Patronage/"

# 2018-2019 train data
URL_train_18_19 = f"{URL_train}Annual%20regional%20train%20station%20entries%202018-19"

# Generate output location and filename
output = f"{output_dir}train_station_18to19.xlsx"

# urllib request
opener=urllib.request.build_opener()
opener.addheaders=[('User-Agent','Mozilla/5.0 (Windows NT 6.1; WOW64) AppleWebKit/537.36 \
    (KHTML, like Gecko) Chrome/36.0.1941.0 Safari/537.36')]
urllib.request.install_opener(opener)

# Urlretrieve the file
urllib.request.urlretrieve(f'{URL_train_18_19}.xlsx', output)

# 2019-2020 train data
URL_train_19_20 = f"{URL_train}Annual%20regional%20train%20station%20entries%202019-20"

# Generate output location and filename
output = f"{output_dir}train_station_19to20.xlsx"

# urllib request
opener=urllib.request.build_opener()
opener.addheaders=[('User-Agent','Mozilla/5.0 (Windows NT 6.1; WOW64) AppleWebKit/537.36 \
    (KHTML, like Gecko) Chrome/36.0.1941.0 Safari/537.36')]
urllib.request.install_opener(opener)

# Urlretrieve the file
urllib.request.urlretrieve(f'{URL_train_19_20}.xlsx', output)


# 2020-2021 train data
URL_train_20_21 = f"{URL_train}Annual%20regional%20train%20station%20entries%202020-21"

# Generate output location and filename
output = f"{output_dir}train_station_20to21.xlsx"

# urllib request
opener=urllib.request.build_opener()
opener.addheaders=[('User-Agent','Mozilla/5.0 (Windows NT 6.1; WOW64) AppleWebKit/537.36 \
    (KHTML, like Gecko) Chrome/36.0.1941.0 Safari/537.36')]
urllib.request.install_opener(opener)

# Urlretrieve the file
urllib.request.urlretrieve(f'{URL_train_20_21}.xlsx', output)

### 4 - Census Data

# URL for Census data in Victoria 2021
URL_census_2021 = 'https://www.abs.gov.au/census/find-census-data/datapacks/download/2021_GCP_all_for_VIC_short-header.zip'
filename = f'{output_dir}census_2021.zip'

# Download the zip file
urlretrieve(URL_census_2021, filename) 

census2021 = zipfile.ZipFile(f'{output_dir}census_2021.zip')

for file in census2021.namelist():
    if file.startswith("2021 Census GCP All Geographies for VIC/POA/"):
        census2021.extract(file, f'{output_dir}/POA')

### 5 - Parks

def get_data():
    url = "https://s3-ap-southeast-2.amazonaws.com/cl-isd-prd-datashare-s3-delivery/Order_AEHQI8.zip"
    response = requests.get(url)
    return url, response.content
 
 
if __name__ == '__main__':
    url, data = get_data()
 
    _tmp_file = tempfile.TemporaryFile()
    print(_tmp_file)
 
    _tmp_file.write(data)
    # _tmp_file.seek(0)
 
    zf = zipfile.ZipFile(_tmp_file, mode='r')
    for names in zf.namelist():
        f = zf.extract(names, '../data/raw')
        print(f)
 
    zf.close()

### 6 - Yearly Income

# URl of yearly income in AUS (2014-2019)
URL_yearly_income = 'https://www.abs.gov.au/statistics/labour/earnings-and-working-conditions/personal-income-australia/2014-15-2018-19/6524055002_DO001.xlsx'

# Generate output location and filename
output = f"{output_dir}Total_income.xlsx"

# Urlretrieve the file
urlretrieve(f'{URL_yearly_income}', output) 

### 7 - LGA Past Rental data (2000-2021)

# URl of past quarter rental dataset (2000-2021)
URL_rent = 'https://www.dhhs.vic.gov.au/quarterly-median-rents-local-government-area-march-quarter-2021-excel'

# Generate output location and filename
output = f"{output_dir}external_rent.xlsx"

# Urlretrieve the file
urlretrieve(f'{URL_rent}', output) 

### 8 - Suburb Past Rental data (2000-2022)

# URl of rental data (2002-2020)
URL_rental = "https://www.dffh.vic.gov.au/sites/default/files/documents/202209/Moving%20annual%20rent%20by%20suburb%20-%20June%20quarter%202022.xlsx"

# Generate output location and filename
output = f"{output_dir}past_rental00_22.xlsx"

# Urlretrieve the file
urlretrieve(f'{URL_rental}', output) 

### 9 - VIC map

If you meet problem about downloading these zip, try "chmod 700 filename".

# download suburb map
import urllib.request
import zipfile
filename='../data/raw/suburb_map.zip'
urllib.request.urlretrieve("https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/access-and-downloads/digital-boundary-files/SAL_2021_AUST_GDA2020_SHP.zip", filename)

# unzip the file
with zipfile.ZipFile(filename,'r') as zip_ref:
    zip_ref.extractall("../data/raw/")

# download postcode map
import urllib.request
import zipfile
filename='../data/raw/population_map.zip'
urllib.request.urlretrieve("https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/access-and-downloads/digital-boundary-files/POA_2021_AUST_GDA2020_SHP.zip", filename)

# unzip the file
with zipfile.ZipFile(filename,'r') as zip_ref:
    zip_ref.extractall("../data/raw/")

# download SA2 map
import urllib.request
import zipfile
filename='../data/raw/SA2_map.zip'
urllib.request.urlretrieve("https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/access-and-downloads/digital-boundary-files/SA2_2021_AUST_SHP_GDA2020.zip", filename)

# unzip the file
with zipfile.ZipFile(filename,'r') as zip_ref:
    zip_ref.extractall("../data/raw/")
