==> notebooks/p1_external_data_preprocess.ipynb <==
# External Data Preprocessing

import pandas as pd
import geopandas as gpd

### 1 - School Location

school_df_1819 = pd.read_csv('../data/raw/school_location_18_19.csv',encoding='cp1252')
school_df_20 = pd.read_csv('../data/raw/school_location_20.csv')
school_df_21 = pd.read_csv('../data/raw/school_location_21.csv',encoding='cp1252')

# select features

school_df_1819 = school_df_1819[['Education_Sector','School_Name','School_Type','Address_Line_1','Address_Town','Address_Postcode','LGA_Name','X','Y']]
school_df_20 = school_df_20[['Education_Sector','School_Name','School_Type','Address_Line_1','Address_Town','Address_Postcode','LGA_Name','X','Y']]
school_df_21 = school_df_21[['Education_Sector','School_Name','School_Type','Address_Line_1','Address_Town','Address_Postcode','LGA_Name','X','Y']]

school_df_1819.isnull().sum().sort_values(ascending=False)

school_df_20.isnull().sum().sort_values(ascending=False)

school_df_21.isnull().sum().sort_values(ascending=False)

# save csv

school_df_1819.to_csv('../data/curated/school_df_1819_preprocessed.csv')
school_df_20.to_csv('../data/curated/school_df_20_preprocessed.csv')
school_df_21.to_csv('../data/curated/school_df_21_preprocessed.csv')

### 3 - Population Data

population_df=pd.read_excel('../data/raw/population.xlsx',sheet_name='Table 1')
population = population_df.loc[8:2461]
pop_drop = population.dropna(axis='columns',how = 'all')
pop_drop.to_csv('../data/raw/population_2.csv', encoding='utf-8')
