==> notebooks/p1_renew_api_data.ipynb <==
# Renew API data

This notebook combines all data from domain and remove the duplicates.

import pandas as pd 

# please note old is the data from the last code run, new is data from this code run
old = pd.read_csv("../rent_data/rent_time2022-09-12", index_col=0)
new = pd.read_csv("../rent_data/rent_time2022-09-13", index_col=0)

def renew_data(old, new):
    """
    function: merge and drop pulication
    para: new data
    """
    renew = pd.concat([old, new])
    renew = renew.reset_index(drop = True).drop_duplicates(subset = ['listingSlug'])
    return renew
renew_data(old, new).to_csv("../rent_data/renew(15).csv")
