==> notebooks/p2_affluence.ipynb <==
# Affluence

import os
import folium
import zipfile
import numpy as np
import pandas as pd
import urllib.request
import geopandas as gpd

# read xlsx file and converted and save to csv file
Total_income=pd.read_excel('../data/raw/Total_income.xlsx',sheet_name='Table 1.4')
Total_income = Total_income.iloc[6:2302,0:27]
income_drop = Total_income.dropna(axis='columns',how = 'all')
income_drop.to_csv('../data/raw/Total_income_2.csv', encoding='utf-8')

# rename the column 
salary_yearly=pd.read_csv('../data/raw/Total_income_2.csv',header=0,names=['SA2','SA2 name', 'Number of earners in 2014-15', 'Number of earners in 2015-16',\
    'Number of earners in 2016-17', 'Number of earners in 2017-18', 'Number of earners in 2018-19', 'Median age of earners in 2014-15',\
        'Median age of earners in 2015-16', 'Median age of earners in 2016-17','Median age of earners in 2017-18', 'Median age of earners in 2018-19',\
            'Total income in 2014-15', 'Total income in 2015-16','Total income in 2016-17', 'Total income in 2017-18','Total income in 2018-19',\
                'Median income in 2014-15','Median income in 2015-16','Median income in 2016-17','Median income in 2017-18','Median income in 2018-19',\
                    'Mean income in 2014-15','Mean income in 2015-16','Mean income in 2016-17','Mean income in 2017-18','Mean income in 2018-19'])

# find the index of Victoria
VIC_total = salary_yearly.loc[salary_yearly['SA2']=='Victoria']

# find the index of Queensland
QUS_total = salary_yearly.loc[salary_yearly['SA2']=='Queensland']

VIC_total

QUS_total

# select the Victoria area 
VIC_salary = salary_yearly.loc[584:1046]
VIC_salary

# Print out the name of the column to be calculated
median_col = VIC_salary.columns.to_list()
columns = []
for i in range(len(median_col)):
    if median_col[i].startswith("Median income"):
        if median_col[i] != "Median income in 2018-19":
            columns.append((median_col[i],median_col[i+1]))
columns

# create function used to calculate the income growth rate
def avg_year(year1, year2):
    result = 0
    if year1!= "np" and year2!="np":
        year1 = int(year1)
        year2 = int(year2)
        if year1 != 0:
            result =  float(year2/year1)
        if year1 == 0:
            result = 0
    return result

# create a function that takes in two columns and spits out the average
df = pd.DataFrame()

for (col_1, col_2) in columns: 
    df[f'{col_1}' + f'{col_2}'] = VIC_salary.apply(lambda x: avg_year(x[f"{col_1}"], x[f'{col_2}']), axis=1)

df['Median Income growth rate'] = (df.sum(axis=1))/len(df.columns)
df

# Add growth rate column in to csv
VIC_salary['Median Income growth rate'] = df['Median Income growth rate']

# List the values of each column
rate = list(df["Median Income growth rate"])
in18_19 = list(VIC_salary["Median income in 2018-19"])
year2014 = list(VIC_salary["Median income in 2014-15"])
year2015 = list(VIC_salary["Median income in 2015-16"])
year2016 = list(VIC_salary["Median income in 2016-17"])
year2017 = list(VIC_salary["Median income in 2017-18"])
year2018 = list(VIC_salary["Median income in 2018-19"])

# Create a new list to store the values of each column
in14 = []
in15 = []
in16 = []
in17 = []
in18 = []
in19_20 = []
in20_21 = []
in21_22 = []
in22_23 = []
in23_24 = []
in24_25 = []
in25_26 = []

# Calculate the weekly income in 2014-2018
for i in range(len(year2014)):
    if year2014[i] != 'np' and year2015[i] != 'np':
        in14.append(round(float(year2014[i])/52,2))
        in15.append(round(float(year2015[i])/52,2))
        in16.append(round(float(year2016[i])/52,2))
        in17.append(round(float(year2017[i])/52,2))
        in18.append(round(float(year2018[i])/52,2))
    # Let all values equal to np be converted to null values.
    else:
        in14.append(np.nan)
        in15.append(np.nan)
        in16.append(np.nan)
        in17.append(np.nan)
        in18.append(np.nan)

# Forecasted annual revenue 2019 to 2025
for i in range(len(rate)):
    if rate[i] != "np" and in18_19[i] != "np":
        
        in19_20.append(round(float(rate[i]) * int(in18_19[i]),2))
  
        in20_21.append(round((rate[i]) * (in19_20[i]),2))

        in21_22.append(round((rate[i]) * (in20_21[i]),2))

        in22_23.append(round((rate[i]) * (in21_22[i]),2))
 
        in23_24.append(round((rate[i]) * (in22_23[i]),2))

        in24_25.append(round((rate[i]) * (in23_24[i]),2))
   
        in25_26.append(round((rate[i]) * (in24_25[i]),2))
    
    # Let all values equal to np be converted to null values.
    else:
        in19_20.append(np.nan)
        in20_21.append(np.nan)
        in21_22.append(np.nan)
        in22_23.append(np.nan)
        in23_24.append(np.nan)
        in24_25.append(np.nan)
        in25_26.append(np.nan)

# Add the yearly income in to original csv
VIC_salary["Median income in 2019-2020"]= in19_20
VIC_salary["Median income in 2020-2021"]= in20_21
VIC_salary["Median income in 2021-2022"]= in21_22
VIC_salary["Median income in 2022-2023"]= in22_23
VIC_salary["Median income in 2023-2024"]= in23_24
VIC_salary["Median income in 2024-2025"]= in24_25
VIC_salary["Median income in 2025-2026"]= in25_26

# Add weekly income in to original csv
VIC_salary["Median weekly income in 2014"]= in14
VIC_salary["Median weekly income in 2015"]= in15
VIC_salary["Median weekly income in 2016"]= in16
VIC_salary["Median weekly income in 2017"]= in17
VIC_salary["Median weekly income in 2018"]= in18

# Drop the null rows in the csv file
VIC_salary.dropna(axis=0, how='any', inplace=True)

# calculate the weekly income in 2019-2025
VIC_salary["Median weekly income in 2019"]= round(VIC_salary["Median income in 2019-2020"]/52,2)
VIC_salary["Median weekly income in 2020"]= round(VIC_salary["Median income in 2020-2021"]/52,2)
VIC_salary["Median weekly income in 2021"]= round(VIC_salary["Median income in 2021-2022"]/52,2)
VIC_salary["Median weekly income in 2022"]= round(VIC_salary["Median income in 2022-2023"]/52,2)
VIC_salary["Median weekly income in 2023"]= round(VIC_salary["Median income in 2023-2024"]/52,2)
VIC_salary["Median weekly income in 2024"]= round(VIC_salary["Median income in 2024-2025"]/52,2)
VIC_salary["Median weekly income in 2025"]= round(VIC_salary["Median income in 2025-2026"]/52,2)

# Select the useful columns of weekly income between 2014 and 2025 and delete the total statistics in Victoria
VIC_salary_weekly = VIC_salary.loc[585:,['SA2 name',\
                          'Median weekly income in 2014','Median weekly income in 2015','Median weekly income in 2016',\
                          'Median weekly income in 2017','Median weekly income in 2018','Median weekly income in 2019',\
                          "Median weekly income in 2020","Median weekly income in 2021",\
                          "Median weekly income in 2022","Median weekly income in 2023",\
                          "Median weekly income in 2024","Median weekly income in 2025"]]
VIC_salary_weekly.to_csv('../data/curated/VIC_weekly_income_SA2.csv')

# Select the useful columns of yearly income between 2014 and 2025 and delete the total statistics in Victoria
VIC_salary_yearly = VIC_salary.loc[585:,['SA2 name',\
                          'Median income in 2014-15','Median income in 2015-16','Median income in 2016-17',\
                          'Median income in 2017-18','Median income in 2018-19',"Median income in 2019-2020",\
                          "Median income in 2020-2021","Median income in 2021-2022",\
                          "Median income in 2022-2023","Median income in 2023-2024",\
                          "Median income in 2024-2025","Median income in 2025-2026"]]
VIC_salary_yearly.to_csv('../data/curated/VIC_yearly_income_SA2.csv')

## Converted SA2 to postcode

# Read the csv file of the converted poa and preprocess
postcode_SA2 = pd.read_csv('../data/raw/postcode_info.csv')
attribute = postcode_SA2.loc[:,['Postcode','SA2 NAME 2016']]
drop_dup = attribute.drop_duplicates(subset=['Postcode'])
rename = drop_dup.rename(columns={'SA2 NAME 2016':'SA2 name'})

# Merge data with weekly income and postcode csv file
VIC_salary_poa_weekly = pd.DataFrame(pd.merge(rename,VIC_salary_weekly, on='SA2 name', how='inner'))

# Merge data with yearly income and postcode csv file
VIC_salary_poa_yearly = pd.DataFrame(pd.merge(rename,VIC_salary_yearly, on='SA2 name', how='inner'))

# Drop the null rows in csv file and save them
VIC_salary_poa_weekly.dropna(axis='rows')
VIC_salary_poa_weekly.to_csv('../data/curated/VIC_weekly_income_poa.csv')
VIC_salary_poa_yearly.dropna(axis='rows')
VIC_salary_poa_yearly.to_csv('../data/curated/VIC_yearly_income_poa.csv')

## Draw SA2 geographic map

sf = gpd.read_file("../data/raw/SA2_2021_AUST_GDA2020.shp")

sf = sf.rename(columns={'SA2_NAME21':'SA2 name'})
sf = sf.loc[sf['STE_NAME21'] == 'Victoria']

sf['geometry'] = sf['geometry'].to_crs("+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs")

result = gpd.GeoDataFrame(pd.merge(VIC_salary[['SA2 name', 'Median Income growth rate']], sf, on='SA2 name', how='inner'))

# create a JSON 
geoJSON = result[['SA2 name','geometry']].drop_duplicates('SA2 name').to_json()

# print the first 300 chars of the json
print(geoJSON[:300])

m = folium.Map(location=[-36, 145.3906], tiles="Stamen Terrain", zoom_start=7)

# refer to the folium documentations on more information on how to plot aggregated data.
c = folium.Choropleth(
    geo_data=geoJSON, # geoJSON 
    name='choropleth', # name of plot
    data=result, # data source
    columns=['SA2 name', 'Median Income growth rate'], # the columns required
    key_on= 'properties.SA2 name', # this is from the geoJSON's properties
    fill_color='YlOrRd', # color scheme
    nan_fill_color='black',
    legend_name='growth rate of median income in 2014-2019'
)

c.add_to(m)

m.save('../plots/income_growth_rate.html')
