==> notebooks/p2_combine.ipynb <==
# Combine

This notebook combines the duration and distance of schools or malls.

#### we ran the code twice. We upload the data directly into the ../distances_and_durations, because it was hard to copy.
#### This is the presentation of data

import pandas as pd
import os
import operator

cur_dir = "../distances_and_durations/school_old/school_distance_duration(final).csv"

# School

csv of Distances and durations from schools to rents add together

# read all the csv

sch1 = pd.read_csv("../distances_and_durations/school_old/distances_durations_school21(1-20).csv")
sch2 = pd.read_csv("../distances_and_durations/school_old/distances_durations_school21(20-39).csv")
sch3 = pd.read_csv("../distances_and_durations/school_old/distances_durations_school21(40-59).csv")
sch4 = pd.read_csv("../distances_and_durations/school_old/distances_durations_school21(60-79).csv")
sch5 = pd.read_csv("../distances_and_durations/school_old/distances_durations_school21(80-119).csv")
sch6 = pd.read_csv("../distances_and_durations/school_old/distances_durations_school21(120-149).csv")
sch7 = pd.read_csv("../distances_and_durations/school_old/distances_durations_school21(150-199).csv")
sch8 = pd.read_csv("../distances_and_durations/school_old/distances_durations_school21(200-249).csv")
sch9 = pd.read_csv("../distances_and_durations/school_old/distances_durations_school21(250-349).csv")
sch10 = pd.read_csv("../distances_and_durations/school_old/distances_durations_school21(350-469).csv")
sch11 = pd.read_csv("../distances_and_durations/school_old/distances_durations_school21(470-519).csv")
sch12 = pd.read_csv("../distances_and_durations/school_old/distances_durations_school21(520-599).csv")
sch13 = pd.read_csv("../distances_and_durations/school_old/distances_durations_school21(600-699).csv")
sch14 = pd.read_csv("../distances_and_durations/school_old/distances_durations_school21(700-819).csv")
sch15 = pd.read_csv("../distances_and_durations/school_old/distances_durations_school21(820-939).csv")
sch16 = pd.read_csv("../distances_and_durations/school_old/distances_durations_school21(940-1059).csv")
sch17 = pd.read_csv("../distances_and_durations/school_old/distances_durations_school21(1060-1179).csv")
sch18 = pd.read_csv("../distances_and_durations/school_old/distances_durations_school21(1180-).csv")

# the output csv
school_dis21_dir = '../distances_and_durations/school_old/school_distance_duration(final).csv' 

# add them together
df = pd.concat([sch1,sch2])
df = pd.concat([df, sch3])
df = pd.concat([df, sch4])
df = pd.concat([df, sch5])
df = pd.concat([df, sch6])
df = pd.concat([df, sch7])
df = pd.concat([df, sch8])
df = pd.concat([df, sch9])
df = pd.concat([df, sch10])
df = pd.concat([df, sch11])
df = pd.concat([df, sch12])
df = pd.concat([df, sch13])
df = pd.concat([df, sch14])
df = pd.concat([df, sch15])
df = pd.concat([df, sch16])
df = pd.concat([df, sch17])
df = pd.concat([df, sch18]).reset_index(drop=True)



df

# add rantal house's postcode and price
rent = pd.read_csv("../rent_data/renew(final).csv")
new_data = df
postcode = []
for i in range(len(new_data)):
    postcode.append(new_data['rent_address'][i][-13:-9])
new_data['postcode'] = postcode
price =[]
for rent_a in new_data['rent_address']:
    price.append(rent[rent['listingSlug'] == rent_a]['price'].tolist()[0])
new_data['price'] = price
new_data = new_data.sort_values(by=["postcode"],ascending=True).reset_index(drop=True)

rent= pd.read_csv("../rent_data/renew(final).csv", index_col=0)
rent_c = pd.read_csv("../rent_data/rent_cleaned.csv", index_col=0)
slice_lable = (
    rent[['listingSlug']].apply(tuple, axis=1)
    .isin(rent_c[['listingSlug']].apply(tuple, axis=1)
          .to_list()
         )
)

rent[~slice_lable]

index = []
for i in range(len(rent[~slice_lable]['listingSlug'])):
    index.append(new_data[new_data['rent_address'] == rent[~slice_lable]['listingSlug'].tolist()[i]].index.tolist()[0])
new_data = new_data.drop(index = index, axis=0)

new_data.to_csv(school_dis21_dir)

# Mall

csv of Distances and durations from malls to rents add together

# list the csvs in the specify folder
mall_list = [x for x in os.listdir("../distances_and_durations/mall_old")]

# go through all the csv in that folder and add them together
df = pd.DataFrame()
for csv_file in mall_list:
    mall_df = pd.read_csv(f"../distances_and_durations/mall_old/{csv_file}")
    df_data = pd.DataFrame(mall_df)
    df = pd.concat([df,df_data])
df = df.reset_index(drop=True)

df

# add rantal house's postcode and price
rent = pd.read_csv("../rent_data/renew(final).csv")
new_data = df
postcode = []
for i in range(len(new_data)):
    postcode.append(new_data['rent_address'][i][-13:-9])
new_data['postcode'] = postcode
price =[]
for rent_a in new_data['rent_address']:
    price.append(rent[rent['listingSlug'] == rent_a]['price'].tolist()[0])
new_data['price'] = price
new_data = new_data.sort_values(by=["postcode"],ascending=True).reset_index(drop=True)

rent= pd.read_csv("../rent_data/renew(final).csv", index_col=0)
rent_c = pd.read_csv("../rent_data/rent_cleaned.csv", index_col=0)
slice_lable = (
    rent[['listingSlug']].apply(tuple, axis=1)
    .isin(rent_c[['listingSlug']].apply(tuple, axis=1)
          .to_list()
         )
)

rent[~slice_lable]

index = []
for i in range(len(rent[~slice_lable]['listingSlug'])):
    index.append(new_data[new_data['rent_address'] == rent[~slice_lable]['listingSlug'].tolist()[i]].index.tolist()[0])
new_data = new_data.drop(index = index, axis=0)

new_data.to_csv('../distances_and_durations/mall_old/mall_distances_durations(final).csv',index = True)
