==> notebooks/p2_population_forecast.ipynb <==
# Population Forcast

import folium
import zipfile
import urllib.request
import pandas as pd
import geopandas as gpd
import numpy as np

# Read csv and rename the columns
population=pd.read_csv('../data/raw/population_2.csv',header=0,names=['S/T code','S/T name', 'GCCSA code', 'GCCSA name',\
                                                                      'SA4 code', 'SA4 name', 'SA3 code', 'SA3 name', 'SA2 code',\
                                                                      'SA2 name', '2001','2002', '2003','2004', '2005','2006',
                                                                      '2007', '2008', '2009', '2010', '2011', '2012', '2013', '2014', '2015',\
                                                                      '2016','2017','2018', '2019','2020','2021','change 2011-2021 number','change 2011-2021 %',\
                                                                      'Area km2','population density 2021(persons/km2)'])

# Select the region of SA2 is Victoria
VIC_population = population.loc[population['S/T name']=='Victoria']
select_population = VIC_population.loc[:, ['S/T name', 'SA2 name', '2001','2002', '2003','2004', '2005','2006',\
                                           '2007', '2008', '2009', '2010', '2011', '2012', '2013', '2014', '2015',\
                                           '2016','2017','2018', '2019','2020','2021','change 2011-2021 number','change 2011-2021 %',\
                                           'Area km2','population density 2021(persons/km2)']] 

VIC_population

# Print out the name of the column to be calculated
col = select_population.columns.to_list()
columns = []
for i in range(len(col)):
    if col[i].startswith("20"):
        if col[i] != "2021":
            columns.append((col[i],col[i+1]))

# create function used to calculate the population growth rate in 20 years
def avg_year(year1, year2):
    result = 0
    if year1!=0:
       result =  year2/year1
    if year1 == 0:
        result = np.nan
    return result

# create a function that takes in two columns and spits out the average
df = pd.DataFrame()

for (col_1, col_2) in columns: 
    df['avg' + f'{col_1}' + f'{col_2}'] = select_population.apply(lambda x: avg_year(x[f"{col_1}"], x[f'{col_2}']), axis=1)

df['average change %'] = (df.sum(axis=1))/len(df.columns)
df

# Add new column in to csv
select_population['population growth rate'] = df['average change %']
select_population

# Predicted population numbers for each of the three years.
select_population["2022"]=(select_population["2021"]*select_population['population growth rate']).round()
select_population["2023"]=(select_population["2022"]*select_population['population growth rate']).round()
select_population["2024"]=(select_population["2023"]*select_population['population growth rate']).round()
select_population["2025"]=(select_population["2024"]*select_population['population growth rate']).round()

# Calcuolate the population density between 2014 and 2025
select_population["population density 2014"]=(select_population["2014"]/select_population["Area km2"]).round()
select_population["population density 2015"]=(select_population["2015"]/select_population["Area km2"]).round()
select_population["population density 2016"]=(select_population["2016"]/select_population["Area km2"]).round()
select_population["population density 2017"]=(select_population["2017"]/select_population["Area km2"]).round()
select_population["population density 2018"]=(select_population["2018"]/select_population["Area km2"]).round()
select_population["population density 2019"]=(select_population["2019"]/select_population["Area km2"]).round()
select_population["population density 2020"]=(select_population["2020"]/select_population["Area km2"]).round()
select_population["population density 2021"]=(select_population["2021"]/select_population["Area km2"]).round()
select_population["population density 2022"]=(select_population["2022"]/select_population["Area km2"]).round()
select_population["population density 2023"]=(select_population["2023"]/select_population["Area km2"]).round()
select_population["population density 2024"]=(select_population["2024"]/select_population["Area km2"]).round()
select_population["population density 2025"]=(select_population["2025"]/select_population["Area km2"]).round()

# Save csv
select_population.to_csv('../data/curated/population_SA2.csv')

# Read the csv file of the converted poa and preprocess
postcode_SA2 = pd.read_csv('../data/raw/postcode_info.csv')
attribute = postcode_SA2.loc[:,['Postcode','SA2 NAME 2016','Latitude','Longitude']]
drop_dup = attribute.drop_duplicates(subset=['Postcode'])
rename = drop_dup.rename(columns={'SA2 NAME 2016':'SA2 name'})

# merge two csv files to convert SA2 to poa
merge_data = pd.DataFrame(pd.merge(rename,select_population,on='SA2 name', how='inner'))

merge_data

# save csv file that is converted poa and drop blank rows
merge_data.dropna(axis='rows')
merge_data.to_csv('../data/curated/population_postcode.csv')

## Draw Geographic Map about population growth rate with SA2 region

sf = gpd.read_file("../data/raw/SA2_2021_AUST_GDA2020.shp")

sf = sf.rename(columns={'SA2_NAME21':'SA2 name'})
sf = sf.loc[sf['STE_NAME21'] == 'Victoria']

sf['geometry'] = sf['geometry'].to_crs("+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs")

# Merge two dataframe and save as Geodataframe
result = gpd.GeoDataFrame(pd.merge(select_population[['SA2 name', 'population growth rate']], sf, on='SA2 name', how='inner'))

# create a JSON 
geoJSON = result[['SA2 name','geometry']].drop_duplicates('SA2 name').to_json()

# print the first 300 chars of the json
print(geoJSON[:300])

m = folium.Map(location=[-36, 145.3906], tiles="Stamen Terrain", zoom_start=7)

# refer to the folium documentations on more information on how to plot aggregated data
# SA2 level data
c = folium.Choropleth(
    geo_data=geoJSON, # geoJSON 
    name='choropleth', # name of plot
    data=result, # data source
    columns=['SA2 name','population growth rate'], # the columns required
    key_on= 'properties.SA2 name', # this is from the geoJSON's properties
    fill_color='YlOrRd', # color scheme
    nan_fill_color='black',
    legend_name='population growth rate in 20 years'
)

c.add_to(m)
m.save('../plots/Population growth rate map.html')
