==> notebooks/p3_affordable.ipynb <==
# Affordable

import os
import pandas as pd
import numpy as np

# read yearly income csv
Yearly_income=pd.read_csv('../data/curated/VIC_yearly_income_poa.csv')

# select the column we need
Yearly_income_poa = pd.DataFrame(Yearly_income, columns=["Postcode", "Median income in 2021-2022"])

# preprocess past rent data
External_rent= pd.read_excel('../data/raw/external_rent.xlsx',sheet_name='All Properties')
External_rent = External_rent.iloc[3:,[1,-1]]
External_rent = External_rent.dropna(axis='rows',how = 'all')

# rename column name
External_rent = External_rent.rename(columns={'Unnamed: 177':'Median rent price in 2021','Unnamed: 1':'LGA'})

External_rent

# Use the weekly rent price to get the total annual rent price
External_rent['Median Yearly rent price in 2021'] = External_rent['Median rent price in 2021']*52

# Read the csv file of the converted poa and preprocess
postcode_LGA = pd.read_csv('../data/raw/postcode_info.csv')
attribute = postcode_LGA.loc[:,['Postcode','LGA Region']]
drop_dup = attribute.drop_duplicates(subset=['Postcode'])
rename = drop_dup.rename(columns={'LGA Region':'LGA'})

# merge two datasets by LGA
External_rent_poa = pd.DataFrame(pd.merge(rename,External_rent, on='LGA', how='inner'))
External_rent_poa.dropna(axis='rows')

# select the columns we need.
External_rent_poa = External_rent_poa.iloc[:,[0,3]]

# merge yearly income and yearly rent price by poa
rent_income = pd.DataFrame(pd.merge(External_rent_poa,Yearly_income_poa, on='Postcode', how='inner'))

# calculate affordable index by yearly income /yearly rent
rent_income['affordable index'] = rent_income['Median income in 2021-2022']/rent_income['Median Yearly rent price in 2021']

# save csv
rent_income.to_csv("../data/curated/rent_income21.csv")

# Sort by affordable index
top_index = rent_income.sort_values(by='affordable index', ascending=False)

# find the top 10 area with affordable index
top_index.iloc[:10,:]
