==> notebooks/p3_liveable.ipynb <==
# Liveable

import pandas as pd
import folium
import geopandas as gpd
from folium.plugins import MarkerCluster
from folium.plugins import HeatMap

school = pd.read_csv("../data/curated/school_df_21_preprocessed.csv")
mall = pd.read_csv("../data/raw/mall_info.csv")
train = pd.read_excel("../data/raw/train_station_20to21.xlsx")
postcode_SA2 = pd.read_csv("../data/raw/postcode_info.csv")
population = pd.read_csv("../data/curated/population_postcode.csv")
rent_income = pd.read_csv("../data/curated/rent_income21.csv")
park = pd.read_csv("../data/curated/park_coord.csv")

# School

school.head()

# each point of the school contains the information of it's latitude, longitude and it's type.

sch_lat = school["Y"].to_list()
sch_lng = school["X"].to_list()
sch_type = school["School_Type"].to_list()

sch_point = []
for i in range(len(sch_lat)):
    sch_point.append((sch_lat[i], sch_lng[i], sch_type[i]))

# Mall

mall.head()

# each point of the mall contains the information of it's latitude, longitude and the number of stores.

mall_lat = mall["Latitude"].to_list()
mall_lng = mall["Longitude"].to_list()
mall_store = mall["Stores"].to_list()

mall_point = []
for i in range(len(mall_lat)):
    mall_point.append((mall_lat[i], mall_lng[i], mall_store[i]))

# Train stop

train.head()

# each point of the train stop contains the information of it's latitude, longitude and stop ID.

train_lat = train["Stop_lat"].to_list()
train_lng = train["Stop_long"].to_list()
train_id = train["Stop_ID"].to_list()

train_point = []
for i in range(len(train_lat)):
    train_point.append((train_lat[i], train_lng[i], train_id[i]))

# Population

population.head()

# each point of the population (for each postcode) contains the information of \
    # it's latitude, longitude and growth rate from 2011 to 2021.

pop_lat = population["Latitude"].to_list()
pop_lng = population["Longitude"].to_list()
pop11_21 = population["2021"].to_list()

pop_point = []
for i in range(len(pop_lat)):
    pop_point.append((pop_lat[i], pop_lng[i], pop11_21[i]))

# Rent and Income

rent_income.head()

# exchange the type for Postcode to string for further mergeing.

rent_income["Postcode"] = rent_income["Postcode"].astype(str)

# only take the attributes Postcode, Longitude and Latitude from the given csv.

postcode = pd.DataFrame(postcode_SA2, columns=["Postcode", "Longitude", "Latitude"])
postcode

# Park

park.head()

# take the attributes park name, Longitude and Latitude from the given csv.

park_coor = park["coordinates"].to_list()
park_name = park["park_name"].to_list()

p_coor = []
for i in park_coor:
    i = i.replace("(","")
    i = i.replace(")", "")
    p_coor.append(i.split(','))

park_point = []
for j in range(len(p_coor)):
    park_point.append((float(p_coor[j][1]), float(p_coor[j][0]), park_name[j]))

# Map

# read the shp file for postcode map for Australia and rename the attribute

sf = gpd.read_file("../data/raw/POA_2021_AUST_GDA2020.shp")
sf = sf.rename(columns={'POA_CODE21':'Postcode'})

sf.head()

# only take the rows for Victoria state

postcode_SA2["Postcode"] = postcode_SA2["Postcode"].astype(str)

poa_vic = postcode_SA2["Postcode"].to_list()
poa_vic = set(poa_vic)

sf_new = pd.DataFrame()
for poa in poa_vic:
    df = sf[sf["Postcode"] == str(poa)]
    sf_new = pd.concat([sf_new, df])

sf_new.head()

sf_new['geometry'] = sf_new['geometry'].to_crs("+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs")

# json file for postcode

geoJSON = sf_new[['Postcode','geometry']].drop_duplicates('Postcode').to_json()

# json file for affordable index on Postcode

re_in = gpd.GeoDataFrame(pd.merge(rent_income[['Postcode', 'affordable index']], sf, on='Postcode', how='inner'))
re_inGeo = re_in[['Postcode','geometry']].drop_duplicates('Postcode').to_json()
print(re_inGeo[:300])

# create layers for different groups

poagroup = folium.FeatureGroup(name='Postcode_Layer', control=True)

popgroup = folium.FeatureGroup(name='Population_Layer', control=True)

schgroup = folium.FeatureGroup(name='School_Layer', control=True)

mallgroup = folium.FeatureGroup(name='Mall_Layer', control=True)

traingroup = folium.FeatureGroup(name='Train_Layer', control=True)

parkgroup = folium.FeatureGroup(name='Park_Layer', control=True)


# add the data to the layers and then to the map

m = folium.Map(location=[-36, 145.3906], tiles="Stamen Terrain", zoom_start=7)

# Postcode layer
folium.GeoJson(
    geoJSON,
    style_function=lambda feature: {
        'fillColor': '#9ad7d2',
        'color': 'black',
        'weight': 2,
        'dashArray': '5, 5'
    },
    popup=folium.GeoJsonPopup(["Postcode"])
).add_to(poagroup)
m.add_child(poagroup)

# population layer
HeatMap(pop_point).add_to(popgroup)
m.add_child(popgroup)

# affordable index layer
affgroup = folium.Choropleth(
    geo_data=re_inGeo, # geoJSON 
    name='Affordable_Layer', # name of plot
    data=re_in, # data source
    columns=['Postcode','affordable index'], # the columns required
    key_on= 'properties.Postcode', # this is from the geoJSON's properties
    fill_color='YlOrRd', # color scheme
    nan_fill_color='black',
    legend_name='Affordable Index'
)
affgroup.add_to(m)

# school layer
# create mark cluster
marker_cluster_sch = MarkerCluster(name="School_Layer").add_to(schgroup)

# add data point to the mark cluster
for (lat, lng, types) in sch_point:
    folium.Marker(
        location=[lat, lng],
        icon=folium.Icon(icon='university',color='black',icon_color='white',angle=45),
        popup=types
    ).add_to(marker_cluster_sch)

# add marker_cluster to map
m.add_child(marker_cluster_sch)


# mall layer
# create mark cluster
marker_cluster_mall = MarkerCluster(name="Mall_Layer").add_to(mallgroup)

# add data point to the mark cluster
for (lat, lng, stores) in mall_point:
    folium.Marker(
        location=[lat, lng],
        icon=folium.Icon(icon='gift',color='orange',icon_color='white',prefix="fa"),
        popup=stores
    ).add_to(marker_cluster_mall)
    
m.add_child(marker_cluster_mall)
 
 
# train stop layer
# create mark cluster           
marker_cluster_train = MarkerCluster(name="Train_Layer").add_to(traingroup)

# add data point to the mark cluster
for (lat, lng, id) in train_point:
    folium.Marker(
        location=[lat, lng],
        icon=folium.Icon(icon='train',color='green',icon_color='white',prefix="fa"),
        popup=id
    ).add_to(marker_cluster_train)
    
m.add_child(marker_cluster_train)


# park layer
# create mark cluster           
marker_cluster_park = MarkerCluster(name="Park_Layer").add_to(parkgroup)

# add data point to the mark cluster
for (lat, lng, name) in park_point:
    folium.Marker(
        location=[lat, lng],
        icon=folium.Icon(icon='train',color='blue',icon_color='white',prefix="fa"),
        popup=name
    ).add_to(marker_cluster_park)
    
m.add_child(marker_cluster_park)

# add controller to the map
folium.LayerControl().add_to(m)

# save the map to html
m.save('../plots/liveable_affordable_poa.html')
